CO2 exploration notebook with some prior results
Results are the Rasmussen kernel with adding the components bit by bit and checking how much they improve the model. This is done on CO2 Data with Z-Scaled X values. Next up should be the exact same procedure as in the Rasmussen i.e. just subtracting the mean of the Y axis